Refactored out a component for resuse by NewPaMessagePage.
