+++ assets/js/components/Dashboard/NewPaMessage/StationsAndZones/SelectedStationsSummary.tsx
@@ -1,12 +1,7 @@
-import { Place } from "Models/place";
-import { Screen } from "Models/screen";
 import React from "react";
-import { GeoAltFill, X } from "react-bootstrap-icons";
-import fp from "lodash/fp";
-import { SILVER_LINE_ROUTES } from "Constants/constants";
-import { classWithModifier } from "../../../../util";
-import pluralize from "pluralize";
-import { usePlacesWithSelectedScreens } from "../hooks";
+import { Place } from "Models/place";
+import { GeoAltFill } from "react-bootstrap-icons";
+import SelectedSignsByRouteTags from "../SelectedSignsByRouteTags";
 
 interface Props {
   value: string[];
@@ -15,45 +10,12 @@ interface Props {
   busRoutes: string[];
 }
 
-const SelectedGroupTag = ({
-  numPlaces,
-  routeId,
-  onClick,
-}: {
-  numPlaces: number;
-  routeId: string;
-  onClick: () => void;
-}) => {
-  if (numPlaces === 0) return null;
-  return (
-    <div
-      className={classWithModifier("selected-group-tag", routeId.toLowerCase())}
-    >
-      {routeId}: {pluralize("Station", numPlaces, true)}
-      <button className={routeId.toLowerCase()} onClick={onClick}>
-        <X />
-      </button>
-    </div>
-  );
-};
-
 const SelectedStationsSummary = ({
   value,
   places,
   busRoutes,
   onChange,
 }: Props) => {
-  const placesWithSelectedScreens = usePlacesWithSelectedScreens(places, value);
-  const removeSelectedScreens = (filterFn: (screen: Screen) => boolean) => {
-    const selectedScreens = fp.flow(
-      fp.flatMap((place: Place) => place.screens),
-      fp.filter(filterFn),
-      fp.map((screen) => screen.id),
-    )(placesWithSelectedScreens);
-
-    onChange(fp.without(selectedScreens, value));
-  };
-
   return (
     <div className="selected-stations-summary">
       <GeoAltFill className="geo-alt-fill-icon" width={12} />
@@ -61,65 +23,12 @@ const SelectedStationsSummary = ({
       {value.length === 0 ? (
         <span>None</span>
       ) : (
-        <>
-          {["Green", "Red", "Orange", "Blue", "Mattapan"].map((routeId) => (
-            <SelectedGroupTag
-              key={routeId}
-              numPlaces={
-                placesWithSelectedScreens.filter((place) =>
-                  place.screens.some((screen) =>
-                    screen.route_ids?.some((r) => r.startsWith(routeId)),
-                  ),
-                ).length
-              }
-              routeId={routeId}
-              onClick={() => {
-                removeSelectedScreens(
-                  (screen) =>
-                    screen.route_ids?.some((r) => r.startsWith(routeId)) ??
-                    false,
-                );
-              }}
-            />
-          ))}
-          <SelectedGroupTag
-            numPlaces={
-              placesWithSelectedScreens.filter((place) =>
-                place.screens.some((screen) =>
-                  screen.route_ids?.some((routeId) =>
-                    SILVER_LINE_ROUTES.includes(routeId),
-                  ),
-                ),
-              ).length
-            }
-            routeId="Silver"
-            onClick={() => {
-              removeSelectedScreens(
-                (screen) =>
-                  fp.intersection(SILVER_LINE_ROUTES, screen.route_ids).length >
-                  0,
-              );
-            }}
-          />
-          <SelectedGroupTag
-            numPlaces={
-              placesWithSelectedScreens.filter((place) =>
-                place.screens.some((screen) =>
-                  screen.route_ids?.some((routeId) =>
-                    busRoutes.includes(routeId),
-                  ),
-                ),
-              ).length
-            }
-            routeId="Bus"
-            onClick={() => {
-              removeSelectedScreens(
-                (screen) =>
-                  fp.intersection(busRoutes, screen.route_ids).length > 0,
-              );
-            }}
-          />
-        </>
+        <SelectedSignsByRouteTags
+          places={places}
+          value={value}
+          onChange={onChange}
+          busRoutes={busRoutes}
+        />
       )}
     </div>
   );
